sentiment_service/requirements.txt
kafka-python
pymongo
vaderSentiment
textblob
